[rust] Make fixed size arrays work with non-Copy types
Fixed size arrays pose challenges for encoding and decoding. The old
code chose to be safe and efficient, but limited (requiring a Copy and
Default bound), so, for example, a fixed-size array of handles was not
possible, and would be a compile error.

The new implementation is unsafe and less efficient, but more
flexible.  Further optimizations are of course possible.

Change-Id: I2df4d593b9f5f90b76acdc9e9de35a5113d0132d